refactor(gateway): verdict helpers live in run_shutdown; keep the early failure return; dedupe test setup
- _exit_with_failure_verdict / _resolve_gateway_exit_verdict move out of the
  run.py facade into run_shutdown.py, which already owns _restart_via_service
  and GATEWAY_SERVICE_RESTART_EXIT_CODE (no alias import needed).
- The running-shutdown tail keeps its early `return False` on a failure verdict,
  as on main, so a failure exit does not first drain cron/MCP; the helper still
  re-checks it for the startup-abort path.
- The three start_gateway tests share one _patch_aborted_startup helper.
